Move all types to single file
